Keep BunchTuneAnalysis4D in same file as BunchTuneAnalysis
