+++ src/meson.build
@@ -80,11 +80,9 @@ sources = files([
 	'orbit/Bunch.cc',
 	'orbit/SyncPart.cc',
 	'orbit/BunchDiagnostics/wrap_bunch_tune_analysis.cc',
-	'orbit/BunchDiagnostics/wrap_bunch_tune_analysis_4d.cc',
 	'orbit/BunchDiagnostics/wrap_bunch_twiss_analysis.cc',
 	'orbit/BunchDiagnostics/BunchTwissAnalysis.cc',
 	'orbit/BunchDiagnostics/BunchTuneAnalysis.cc',
-	'orbit/BunchDiagnostics/BunchTuneAnalysis4D.cc',
 	'orbit/FieldTracker/wrap_fieldtracker.cc',
 	'orbit/FieldTracker/FieldTracker.cc',
 	'orbit/MaterialInteractions/wrap_foil.cc',
