--- src/orbit/BunchDiagnostics/wrap_bunch_tune_analysis_4d.hh
@@ -1,18 +1,0 @@
-#ifndef WRAP_BUNCH_TUNE_ANALYSIS_4D_HH_
-#define WRAP_BUNCH_TUNE_ANALYSIS_4D_HH_
-
-#include "Python.h"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-  namespace wrap_bunch_tune_analysis_4d{
-    void initbunchtuneanalysis4d(PyObject* module);
-  }
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif /*WRAP_BUNCH_TUNE_ANALYSIS_4D_HH_*/
